refactor: rename package xelon to driver
